pyproject.toml
[project]
name = "scraper-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "beautifulsoup4==4.12.3",
    "blinker==1.9.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "distro==1.9.0",
    "extruct==0.18.0",
    "flask==3.0.3",
    "flask-cors==5.0.0",
    "h11==0.14.0",
    "html-text==0.6.2",
    "html5lib==1.1",
    "httpcore==1.0.7",
    "httpx==0.27.2",
    "idna==3.10",
    "isodate==0.7.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "jiter==0.7.1",
    "jstyleson==0.0.2",
    "lxml==5.3.0",
    "lxml-html-clean==0.3.1",
    "markupsafe==3.0.2",
    "mf2py==2.0.1",
    "openai==1.54.5",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pyparsing==3.2.0",
    "pyrdfa3==3.6.4",
    "python-dotenv==1.0.1",
    "rdflib==7.1.1",
    "recipe-scrapers==15.2.1",
    "requests==2.32.3",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "tqdm==4.67.0",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "w3lib==2.2.1",
    "webencodings==0.5.1",
    "werkzeug==3.1.3",
]
